feat!: migrate to TypeScript
BREAKING CHANGE: Named export and minimum Node.js version is now v22.
