--- .github/workflows/nodejs.yml
@@ -1,28 +1,0 @@
-name: nodejs
-on: [pull_request, push]
-permissions:
-  contents: read
-jobs:
-  build:
-    runs-on: ${{ matrix.os }}
-    strategy:
-      matrix:
-        node-version: [18, 20]
-        os: [ubuntu-latest]
-    steps:
-    - name: Node.js ${{ matrix.node-version }}
-      uses: actions/setup-node@48b55a011bda9f5d6aeb4c2d9c7362e8dae4041e # v6.4.0
-      with:
-        node-version: ${{ matrix.node-version }}
-    - name: Update npm
-      run: |
-        npm install -g npm
-        npm --version
-    - uses: actions/checkout@df4cb1c069e1874edd31b4311f1884172cec0e10 # v6.0.3
-    - name: Install dependencies
-      uses: bahmutov/npm-install@20216767ca67f0f7b4d095dc5859c5700a6581cb # v1.12.1
-      with:
-        useLockFile: false
-    - run: npm ls
-    - name: Test
-      run: npm test
