+++ .npmrc
@@ -1,2 +1,5 @@
-package-lock = false
+enable-pre-post-scripts = true
 provenance = true
+save-exact = true
+shell-emulator = true
+update-notifier = false
