--- CHANGELOG.md
@@ -1,12 +1,0 @@
-## [2.0.0](https://github.com/kenany/clickerheroes-format/compare/1.0.7...2.0.0) (2023-09-25)
-
-
-### ⚠ BREAKING CHANGES
-
-* Node.js v14 and v16 are no longer supported.
-* Node.js v12 is no longer supported.
-
-### Features
-
-* drop Node.js v12 support ([1d14156](https://github.com/kenany/clickerheroes-format/commit/1d1415686f283a689a9be438ef00ab10640646a1))
-* drop Node.js v14 and v16 ([92518ba](https://github.com/kenany/clickerheroes-format/commit/92518ba71f6e9473cf338024c5a5e559f315afac))
