+++ LICENSE.txt
@@ -1,4 +1,4 @@
-Copyright 2014–2023 Kenan Yildirim <https://kenany.me/>
+Copyright 2014–2026 Kenan Yildirim <https://kenany.me/>
 
 Permission is hereby granted, free of charge, to any person obtaining a copy of
 this software and associated documentation files (the "Software"), to deal in
