+++ package.json
@@ -1,41 +1,66 @@
 {
   "name": "clickerheroes-format",
-  "version": "2.0.0",
+  "version": "0.0.0-semantic-release",
   "description": "Format large numbers like ClickerHeroes does",
   "keywords": [
-    "clickerheroes"
+    "clickerheroes",
+    "format",
+    "number"
   ],
-  "repository": "github:kenany/clickerheroes-format",
   "license": "MIT",
   "author": "Kenan Yildirim <kenan@kenany.me> (https://kenany.me/)",
-  "main": "index.js",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/kenany/clickerheroes-format.git"
+  },
+  "publishConfig": {
+    "access": "public"
+  },
+  "type": "commonjs",
+  "types": "dist/index.d.cts",
+  "main": "dist/index.cjs",
+  "exports": {
+    ".": {
+      "import": "./dist/index.mjs",
+      "require": "./dist/index.cjs"
+    },
+    "./*.mjs": {
+      "default": "./dist/*.mjs"
+    },
+    "./*.cjs": {
+      "default": "./dist/*.cjs"
+    },
+    "./*": {
+      "import": "./dist/*.mjs",
+      "require": "./dist/*.cjs"
+    }
+  },
   "files": [
-    "index.js",
+    "dist",
     "LICENSE.txt"
   ],
-  "directories": {
-    "test": "test"
-  },
   "engines": {
-    "node": "18 || >=20"
+    "node": "22 || 24 || >=26"
   },
   "scripts": {
-    "test": "tape test/index.js",
-    "posttest": "npm run -s lint",
-    "lint": "eslint .",
-    "release": "semantic-release"
+    "build": "tsdown && tsc --noEmit",
+    "lint": "biome check",
+    "test": "vitest run",
+    "publint": "publint --strict",
+    "prepare": "pnpm build"
   },
   "dependencies": {
     "humanize-number": "^0.0.2"
   },
   "devDependencies": {
-    "@kenan/eslint-config": "11.1.18",
-    "@semantic-release/changelog": "6.0.3",
-    "@semantic-release/git": "10.0.1",
+    "@biomejs/biome": "2.4.15",
+    "@containerbase/semantic-release-pnpm": "1.4.1",
+    "@kenan/biome-config": "1.0.6",
     "conventional-changelog-conventionalcommits": "9.3.1",
-    "eslint": "8.57.1",
-    "lodash.isfunction": "3.0.9",
+    "publint": "0.3.21",
     "semantic-release": "25.0.3",
-    "tape": "5.9.0"
+    "tsdown": "0.22.0",
+    "typescript": "6.0.3",
+    "vitest": "4.1.7"
   }
 }
