+++ src/humanize-number.d.ts
@@ -1,0 +1,3 @@
+declare module 'humanize-number' {
+  export default function humanize(n: number): string;
+}
