+++ vitest.config.mts
@@ -1,0 +1,3 @@
+import { defineConfig } from 'vitest/config';
+
+export default defineConfig({ resolve: { tsconfigPaths: true } });
